Make the fares date helpers private methods
Reverses the earlier move to PlainDate.ts. startOfYear and
getFirstDateAfter are internal to CleanFaresCommand rather than general
date utilities - getFirstDateAfter takes an MMDD restriction month,
which is a fares feed concept - so they become private methods on the
command and PlainDate.ts returns to its previous contents.

CleanFaresCommand still exports only the command class, which was the
point of the review comment.

The nine tests covering them are dropped, since the only public entry
point on the class is run() and that needs a database connection.
