+++ src/cli/CleanFaresCommand.ts
@@ -1,7 +1,6 @@
 
 import {CLICommand} from "./CLICommand";
 import {DatabaseConnection} from "../database/DatabaseConnection";
-import {getFirstDateAfter, startOfYear} from "../gtfs/native/PlainDate";
 
 export class CleanFaresCommand implements CLICommand {
   private readonly queries = [
@@ -77,7 +76,7 @@ export class CleanFaresCommand implements CLICommand {
     const [[current, future]] = await this.db.query<RestrictionDateRow>("SELECT * FROM restriction_date ORDER BY cf_mkr");
 
     await Promise.all(this.restrictionTables.map(
-      t => this.updateRestrictionDatesOnTable(t, startOfYear(current.start_date), startOfYear(future.start_date))
+      t => this.updateRestrictionDatesOnTable(t, this.startOfYear(current.start_date), this.startOfYear(future.start_date))
     ));
 
     console.log("Applied restriction dates");
@@ -89,8 +88,8 @@ export class CleanFaresCommand implements CLICommand {
     const [records] = await this.db.query<RestrictionRow>(`SELECT * FROM ${tableName}`);
     const promises = records.map(record => {
       const earliestDate = record.cf_mkr === 'C' ? current : future;
-      const startDate = getFirstDateAfter(earliestDate, record.date_from);
-      const endDate = startDate && getFirstDateAfter(startDate, record.date_to);
+      const startDate = this.getFirstDateAfter(earliestDate, record.date_from);
+      const endDate = startDate && this.getFirstDateAfter(startDate, record.date_to);
 
       if (!startDate || !endDate || Temporal.PlainDate.compare(startDate, endDate) > 0) {
         console.log(`Invalid dates on ${tableName}: ${record.date_from}, ${record.date_to} after ${earliestDate.toString()}`);
@@ -144,6 +143,33 @@ export class CleanFaresCommand implements CLICommand {
     `);
 
     console.log("Calculated network area restrictions");
+  }
+
+  /**
+   * The 1st of January in the year of the given date
+   */
+  private startOfYear(date: string): Temporal.PlainDate {
+    return Temporal.PlainDate.from(date).with({ month: 1, day: 1 });
+  }
+
+  /**
+   * Given a short form restriction month MMDD this method will return the first instance of that date that occurs
+   * after the given date. For example with a restriction date of 2017-06-01 the earliest date of 0301 is 2018-03-01.
+   *
+   * Returns undefined when the restriction month is not a real date in the resulting year - 0229 outside a leap
+   * year, say. moment returned an invalid date for those and the caller tested isValid(); Temporal rejects them.
+   */
+  private getFirstDateAfter(earliestDate: Temporal.PlainDate, restrictionMonth: string): Temporal.PlainDate | undefined {
+    const month = +restrictionMonth.slice(0, 2);
+    const day = +restrictionMonth.slice(2);
+    const yearOffset = (earliestDate.month > month || (earliestDate.month === month && earliestDate.day > day)) ? 1 : 0;
+
+    try {
+      return Temporal.PlainDate.from({ year: earliestDate.year + yearOffset, month, day }, { overflow: "reject" });
+    }
+    catch {
+      return undefined;
+    }
   }
 
 }
